refactor: re-export the helpers
