+++ components/app/Report/RdReportExtras.vue
@@ -3,7 +3,7 @@ import RdTitle from '~/components/shared/RdTitle.vue'
 import RdUnorderedList from '~/components/shared/RdUnorderedList.vue'
 import RdQAndA from '~/components/shared/RdQAndA.vue'
 
-import intersect from '~/helpers/vue/directives/intersect.js'
+import { intersect } from '~/helpers/vue/directives/index.js'
 
 import {
   setupIntersectionObserver,
