+++ components/helpers/vue/mixins/index.js
@@ -1,0 +1,1 @@
+export { default as scrollDirection } from './scroll-direction.js'
