+++ helpers/vue/directives/index.js
@@ -1,0 +1,1 @@
+export { default as intersect } from './intersect.js'
